remove unused code after refactor
